Reorganize tuple's implementation files underneath agency/detail/tuple
