+++ agency/detail/shape_tuple.hpp
@@ -2,7 +2,7 @@
 
 #include <agency/detail/config.hpp>
 #include <agency/tuple.hpp>
-#include <agency/tuple/detail/arithmetic_tuple_facade.hpp>
+#include <agency/detail/tuple/arithmetic_tuple_facade.hpp>
 #include <agency/detail/type_traits.hpp>
 #include <agency/detail/make_tuple_if_not_scoped.hpp>
